requirements-test.txt
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov
httpx==0.28.1
asyncpg==0.30.0
python-decouple==3.8